Rename assetsvc_utils to utils
