+++ cmd/assetsvc/pkg/utils/postgresql_utils_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package assetsvc_utils
+package utils
 
 import (
 	"database/sql/driver"
