+++ cmd/assetsvc/pkg/utils/utils.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package assetsvc_utils
+package utils
 
 import (
 	"github.com/kubeapps/common/datastore"
